Move QgsDataItem python createChildren crash test to disabled tests
This test was originally working under Qt 5.12, but is broken on newer Qt or sip
versions. The test currently segfaults, as the children created by the python QgsDataCollectionItem
subclass PyQgsDataConnectionItem are immediately garbage collected.

The SIP SIP_VIRTUAL_CATCHER_CODE in qgsdataitem.h is supposed to fix this situation by
adding an extra reference to the returned python objects, but the lines

  // pyItem is given an extra reference which is removed when the C++ instance’s destructor is called.
  sipTransferTo( pyItem, Py_None );

no longer have any effect and the object is still immediately deleted.

Attempted solutions include:
- all combinations of the existing VirtualCatcherCode with the different Factory/TransferBack annotations
- removing the VirtualCatcherCode and replacing with raw Factory/TransferBack annotations
- disabling the python garbage collection of the object entirely with sipTransferTo( pyItem, NULL )

I can't solve this, so for now we have no choice but to disable the test on newer Qt
